Handle non-exported package.json files
node's resolve throws an ERR_PACKAGE_PATH_NOT_EXPORTED error when trying
to resolve the package.json of a package which does not list itself in
its "exports" object.

This works around the issue by handling this case separately, using the
error message to get the path to the file, which is then required
correctly by the existing call to tryRequire().